finish wdio browser test skipping
