+++ tests/specs/test.e2e.js
@@ -1,4 +1,8 @@
-import Matcher from '../../tools/jil/util/browser-matcher'
+import SpecMatcher from '../../tools/browser-matcher/spec-matcher.mjs'
+import { notInternetExplorer } from '../../tools/browser-matcher/common-matchers.mjs'
+
+const notChrome = new SpecMatcher()
+  .exclude('chrome')
 
 describe('newrelic api', () => {
   let testHandle
@@ -11,7 +15,23 @@ describe('newrelic api', () => {
     await testHandle.destroy()
   })
 
-  test(Matcher.withFeature('notInternetExplorer'))('not ie', async () => {
+  // test(notInternetExplorer)('not ie', async () => {
+  //   const url = await testHandle.assetURL('instrumented.html', {
+  //     loader: 'spa'
+  //   })
+
+  //   await Promise.all([
+  //     testHandle.expectRum(),
+  //     browser.url(url)
+  //   ])
+  //   const result = await browser.execute(function () {
+  //     return typeof window.newrelic.addToTrace === 'function'
+  //   })
+
+  //   expect(result).toEqual(true)
+  // })
+
+  test(notChrome)('not chrome', async () => {
     const url = await testHandle.assetURL('instrumented.html', {
       loader: 'spa'
     })
@@ -26,20 +46,4 @@ describe('newrelic api', () => {
 
     expect(result).toEqual(true)
   })
-
-  // test(Matcher.withFeature('workerStackSizeGeneratesError'))('not firefox', async () => {
-  //   const url = await testHandle.assetURL('api/instrumented.html', {
-  //     loader: 'spa'
-  //   })
-
-  //   await Promise.all([
-  //     testHandle.expectRum(),
-  //     browser.url(url)
-  //   ])
-  //   const result = await browser.execute(function () {
-  //     return typeof window.newrelic.addToTrace === 'function'
-  //   })
-
-  //   expect(result).toEqual(true)
-  // })
 })
