+++ tools/browser-matcher/common-matchers.mjs
@@ -1,0 +1,5 @@
+import SpecMatcher from './spec-matcher'
+
+// Does the browser have a reliable 'unload' event callback?
+export const notInternetExplorer = new SpecMatcher()
+  .exclude('ie')
