+++ tools/wdio/plugins/jil-commands.mjs
@@ -3,7 +3,6 @@
  */
 export default class JilCommands {
   async before (capabilities, context, browser) {
-    console.log(global.it)
     browser.addCommand('waitForFeature', async function (feature) {
       const command = `window.NREUM && window.NREUM.activatedFeatures && window.NREUM.activatedFeatures['${feature}']`
       await browser.waitUntil(
